{"types":"C","runs":4}

+++ Problem035.java
@@ -5,8 +5,8 @@ public class Problem035 extends ConsoleProgram {
 	private RandomGenerator rgen = RandomGenerator.getInstance();
 	private static final int NUM_EXPERIMENTS = 1000;
 	public void run(){
-		int sumFlips = 0;
-		int curFlips = 0;
+		double sumFlips = 0;
+		double curFlips = 0;
 		for( int i = 0; i < NUM_EXPERIMENTS; i++){
 			sumFlips ++;
 			if (isHeads()){
